refactor: move cat-list to shared components
