+++ src/components/cat-list/model/types.ts
